Force columns to resize if they are set to autoresize.
10.10 introduces a different behavior with regards to resizing columns
with no defined width. Previously, they would have their width extended
to the edge of the table view. in 10.10, they use the default width
(100) and stay at that width. I've experimented with the different
flags and with inserting very large values for column widths. Neither
of those work. This fix only works when all columns have no defined
width. If one or more columns have a defined width the OS isn't forced
to redraw them as that will use the minWidth for columns with widths.
vanilla doesn't set that and the default is 10, so things go bad
quickly.