Add maxWidth and minWidth keys for columns.
I realized that if a column has a defined width, the column is not
allowed to be resized (by vanilla). This can be used to help with some
of the autoresizing issues. So, always size everything to fit. Also add
minWidth and maxWidth keys for column descriptions so that this can be
customized.