Make RadioGroup work with auto layout.
NSMatrix has been soft deprecated by Apple and it does not work in Auto Layout. Rather than trying to bolt on more `if` conditions to the already complicated RadioGroup, I'm adding HorizontalRadioGroup and VerticalRadioGroup. These are built on the vanilla *StackGroup classes and use the new RadioButton. All of this allows RadioGroup to continue functioning as it has without even possibly breaking existing code.